Shoti
shoti is a secure and fun way to share short videos with your friends. Create and send short videos as reels with ease, ensuring privacy with secure sending. Enjoy a smooth experience with login and signup features, and share your moments in a snap! With ShotShare, you can send and enjoy your videos in style, without the need for a chat feature.